+++ src/macros.rs
@@ -25,7 +25,7 @@
 /// ```
 #[macro_export]
 macro_rules! actions {
-    ( $( $action:expr ),* $(,)? ) => {
-        [ $( ::core::convert::Into::<::std::boxed::Box<dyn $crate::Action>>::into($action) ),* ]
+    ( $( $action:expr ),+ $(,)? ) => {
+        [ $( $crate::IntoBoxedAction::into_boxed_action($action) ),+ ]
     }
 }