+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## Version 0.12.0-dev
 
+- [Add multiple actions with a tuple][104]
+    - Removes `add_many` as `add` can now be used instead
 - [Add cleanup of actions for despawned agents][101]
     - `on_stop`, `on_remove` and `on_drop` in the `Action` trait
     now take `Option<Entity>` for `agent` in order to respond to despawns
     - Replaces all unwraps with logging
     - Adds `on_remove_hook` and `on_remove_trigger` for both `CurrentAction` and `ActionQueue`
 
+[104]: https://github.com/hikikones/bevy-sequential-actions/pull/104
 [101]: https://github.com/hikikones/bevy-sequential-actions/pull/101
 
 ## Version 0.11.0